Update README.md
This Project depicts our work on linear Regression with Car Data. We took the dataset and performed the cleansing techniques on it and post this plotted multiple graphs (Bar , Heatmaps) to depict the relations between various factors Vs. Car selling Price. in third step we performed Linear regression ignoring Car Name and selling price(y fig.) . In the last step we did sentimental analysis on car models vs. the reviews by customer on that model. In this we failed at a step where plotting against 3args vs a blank field is there. I tried solving it by using python attributes of Nan and other but still a scope to see at the data.